started auto-notation-rotation
basically I am tired of having to manually rotate the moves that I am applying to the cube in my head, so I am started a new tab in the visualizer to rotate the moves for me. I commented out the cubevis tab because at the moment there are a few bugs that I need to fix but I don't care enough to put a bunch of time in right now. I think at some point after I finish this part of the project,  I might integrate rotation into the main processing of moves but I'm not sure how much I would really use it so I won't yet.